Made an edit...
Text before edit:
"This example also introduces the 'Insight' concept, sometimes named as 'Intelligence'. It's an sporadic manifestation of Wisdom. It's the natural evolution to it, and what really connects Knowledge and Wisdom."
Text after edit: 
"This example also introduces the 'Insight' concept, sometimes known as 'Intelligence'. It's a sporadic manifestation of Wisdom. Insight is what connects Knowledge and Wisdom."